Add preliminary ARM support and feature detection.
This change detects if the hardware supports an FMA instruction by
issuing it and recovering from the subsequent SIGILL. The 'hasFMA' flag
which is set is then used by the FMA function to issue the hardware
instructions when supported. In order to allow compilation on ARM
targets, the FMOV and FMADD instructions have to be written with WORD
directives.

Additional testing still needs to be done across all toolchain-supported
ARM systems.