[x86,isFMASupported] AND w/ 1<<12 and MOV into CH.
If we are only masking a single bit in CX, we can just move those 8 bits
into CH as long as we address them properly. Instead of shifting down
and AND'ing the first bit, we can just AND the 12th bit, 0-indexed. When
CH is coerced into a bool, it will only be true if it is nonzero.