Add translation of SoftFP's FMA implementation
This change translates Fabrice Bellard's SoftFP implementation from C to
Go. Notably, an F_SIZE of 64 is chosen, and in accordance with IEEE-754,
roundTiesToEven is used as the rounding mode. Benchmarks reveal it to be
on par with MUSL's implementation, which may not warrant the additional
50 lines of code.

    goos: darwin
    goarch: amd64
    pkg: github.com/smasher164/fma
    BenchmarkBSD-4       	20000000	        58.9 ns/op
    BenchmarkMUSL-4      	50000000	        33.8 ns/op
    BenchmarkBellard-4   	50000000	        33.4 ns/op
    PASS
    ok  	github.com/smasher164/fma	4.677s

Updates github.com/smasher164/fma/issues/3.