Add custom software implementation
This change adds a custom software implementation that has been
pulled out of my CL 127458
(https://go-review.googlesource.com/c/go/+/127458).

It has been hand-optimized to benefit from branch prediction, and
outperforms MUSL by ~5ns for RNE. There are still some potential wins
to be made by special-casing cancellation during the effective
subtraction phase, and reordering branches in the compressed shifts.
The shift logic may also be doing too much.

Note: This implementation is written to be convertible into a
branchless one if needed. However the branchless logic tends to be
accompanied by a greater arithmetic instruction count. If constant-time
properties are not desired, it is better to prefer branches into the
fast path.

This implementation is licensed as permissively as possible
(public-domain equivalent), which implies that future languages
and libraries may use this code WITHOUT any attribution, and WITHOUT
including this library's LICENSE.

Assembly has been removed for now, and may be re-added in the future
with a hopefully cleaner organization, probably with x/sys/cpu.